# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PET - Portable Event Telephone API
  version: 1.0.0
  description: REST API for events, extensions, devices, DECT, callbacks and more.
paths:
...
  /api/v1/dect/alerts/:
    get:
      operationId: v1_dect_alerts_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAlertList'
          description: ''
...
  /api/v1/dect/clusters/:
    get:
      operationId: v1_dect_clusters_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSyncClusterList'
          description: ''
...
  /api/v1/dect/coverage/:
    get:
      operationId: v1_dect_coverage_retrieve
      description: Per-RFP handset counts, per-cluster health, totals and weak zones
        (orga/helpdesk).
      parameters:
      - in: query
        name: event
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/dect/handsets/:
    get:
      operationId: v1_dect_handsets_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedHandsetList'
          description: ''
...
  /api/v1/dect/rfps/:
    get:
      operationId: v1_dect_rfps_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRFPList'
          description: ''
...
  /api/v1/dect/sync/:
    post:
      operationId: v1_dect_sync_create
      description: Trigger an immediate OMM sync for one event (orga only).
      parameters:
      - in: query
        name: event
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/devices/{id}/:
    get:
      operationId: v1_devices_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSecret'
          description: ''
    put:
      operationId: v1_devices_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceSecretRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DeviceSecretRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DeviceSecretRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSecret'
          description: ''
    patch:
      operationId: v1_devices_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedDeviceSecretRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedDeviceSecretRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedDeviceSecretRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceSecret'
          description: ''
    delete:
      operationId: v1_devices_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
    Alert:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
          readOnly: true
        severity:
          $ref: '#/components/schemas/SeverityEnum'
        kind:
          type: string
          maxLength: 40
        rfp:
          type: integer
          nullable: true
        rfp_name:
          type: string
          readOnly: true
        message:
          type: string
          maxLength: 300
        created_at:
          type: string
          format: date-time
          readOnly: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        notified:
          type: boolean
      required:
      - created_at
      - event
      - id
      - kind
      - message
      - rfp_name
      - severity
...
    DeviceSecret:
      type: object
      description: Includes SIP password / subscription PIN / softphone QR payloads
        - only for owner/orga.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        event:
          type: string
          readOnly: true
        owner:
          type: string
          title: Nickname
          description: Public handle shown in phonebooks.
          readOnly: true
        type:
          $ref: '#/components/schemas/TypeEdeEnum'
        name:
          type: string
          description: Free-form label, e.g. 'my orange handset'
          maxLength: 80
        state:
          allOf:
          - $ref: '#/components/schemas/State854Enum'
          readOnly: true
        ipei:
          oneOf:
          - type: string
            pattern: ^\d{13}$
            maxLength: 13
          - type: string
            maxLength: 0
        handset_model:
          type: string
          maxLength: 60
        omm_ppn:
          type: string
          readOnly: true
          description: Portable part number in the OMM.
        last_seen_rfp:
          type: integer
          readOnly: true
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        battery_percent:
          type: integer
          readOnly: true
          nullable: true
        rssi:
          type: integer
          readOnly: true
          nullable: true
        sip_username:
          type: string
          readOnly: true
        sip_transport:
          $ref: '#/components/schemas/SipTransportEnum'
        sip_registered_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        mac_address:
          type: string
          maxLength: 17
        extensions:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        sip_password:
          type: string
          maxLength: 128
        subscription_pin:
          type: string
          maxLength: 16
        subscription_pin_expires_at:
          type: string
          format: date-time
          nullable: true
        sip_uri:
          type: string
          readOnly: true
        softphone_links:
          type: string
          readOnly: true
      required:
      - battery_percent
      - created_at
      - event
      - extensions
      - id
      - ipei
      - last_seen_at
      - last_seen_rfp
      - omm_ppn
      - owner
      - rssi
      - sip_registered_at
      - sip_uri
      - sip_username
      - softphone_links
      - state
      - type
    DeviceSecretRequest:
      type: object
      description: Includes SIP password / subscription PIN / softphone QR payloads
        - only for owner/orga.
      properties:
        type:
          $ref: '#/components/schemas/TypeEdeEnum'
        name:
          type: string
          description: Free-form label, e.g. 'my orange handset'
          maxLength: 80
        ipei:
          oneOf:
          - type: string
            pattern: ^\d{13}$
            maxLength: 13
          - type: string
            maxLength: 0
        handset_model:
          type: string
          maxLength: 60
        sip_transport:
          $ref: '#/components/schemas/SipTransportEnum'
        mac_address:
          type: string
          maxLength: 17
        sip_password:
          type: string
          maxLength: 128
        subscription_pin:
          type: string
          maxLength: 16
        subscription_pin_expires_at:
          type: string
          format: date-time
          nullable: true
      required:
      - ipei
      - type
...
    Handset:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        event:
          type: string
          readOnly: true
        owner:
          type: string
          readOnly: true
        name:
          type: string
          description: Free-form label, e.g. 'my orange handset'
          maxLength: 80
        ipei:
          oneOf:
          - type: string
            pattern: ^\d{13}$
            maxLength: 13
          - type: string
            maxLength: 0
        state:
          $ref: '#/components/schemas/State854Enum'
        omm_ppn:
          type: string
          description: Portable part number in the OMM.
          maxLength: 32
        handset_model:
          type: string
          maxLength: 60
        last_seen_rfp:
          type: string
          readOnly: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        battery_percent:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        rssi:
          type: integer
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          format: int64
          nullable: true
        extensions:
          type: string
          readOnly: true
      required:
      - event
      - extensions
      - id
      - ipei
      - last_seen_rfp
      - owner
...
    PaginatedAlertList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
...
    PaginatedHandsetList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Handset'
...
    PaginatedRFPList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/RFP'
...
    PaginatedSyncClusterList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/SyncCluster'
...
    PatchedDeviceSecretRequest:
      type: object
      description: Includes SIP password / subscription PIN / softphone QR payloads
        - only for owner/orga.
      properties:
        type:
          $ref: '#/components/schemas/TypeEdeEnum'
        name:
          type: string
          description: Free-form label, e.g. 'my orange handset'
          maxLength: 80
        ipei:
          oneOf:
          - type: string
            pattern: ^\d{13}$
            maxLength: 13
          - type: string
            maxLength: 0
        handset_model:
          type: string
          maxLength: 60
        sip_transport:
          $ref: '#/components/schemas/SipTransportEnum'
        mac_address:
          type: string
          maxLength: 17
        sip_password:
          type: string
          maxLength: 128
        subscription_pin:
          type: string
          maxLength: 16
        subscription_pin_expires_at:
          type: string
          format: date-time
          nullable: true
...
    RFP:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
          readOnly: true
        omm_id:
          type: string
          maxLength: 32
        name:
          type: string
          maxLength: 80
        mac_address:
          type: string
          maxLength: 17
        ip_address:
          type: string
          nullable: true
        location:
          type: string
          maxLength: 120
        cluster:
          type: string
          readOnly: true
        sync_source:
          type: string
          maxLength: 32
        is_active:
          type: boolean
        connected:
          type: boolean
        synced:
          type: boolean
        status:
          type: string
          readOnly: true
        last_state_change:
          type: string
          format: date-time
          nullable: true
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        active_calls:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        venue_map:
          type: integer
          nullable: true
        pos_x:
          type: number
          format: double
          nullable: true
        pos_y:
          type: number
          format: double
          nullable: true
        handsets:
          type: integer
          readOnly: true
          default: 0
      required:
      - cluster
      - event
      - handsets
      - id
      - name
      - omm_id
      - status
...
    SeverityEnum:
      enum:
      - info
      - warning
      - critical
      type: string
      description: |-
        * `info` - Info
        * `warning` - Warning
        * `critical` - Critical
    SipTransportEnum:
      enum:
      - udp
      - tcp
      - tls
      - wss
      type: string
      description: |-
        * `udp` - UDP
        * `tcp` - TCP
        * `tls` - TLS
        * `wss` - WSS
    State854Enum:
      enum:
      - new
      - pending
      - subscribed
      - offline
      - disabled
      type: string
      description: |-
        * `new` - New
        * `pending` - Pending subscription
        * `subscribed` - Subscribed / registered
        * `offline` - Offline
        * `disabled` - Disabled
...
    SyncCluster:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
          readOnly: true
        cluster_id:
          type: string
          maxLength: 16
        name:
          type: string
          maxLength: 80
        health:
          type: string
          readOnly: true
      required:
      - cluster_id
      - event
      - health
      - id
...
    TypeEdeEnum:
      enum:
      - dect
      - sip
      - webrtc
      - gsm
      - analog
      type: string
      description: |-
        * `dect` - DECT handset
        * `sip` - SIP endpoint
        * `webrtc` - WebRTC softphone
        * `gsm` - GSM handset
        * `analog` - Analog / ATA
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid